+++ src/services/CacheRequestService.php
@@ -285,8 +285,10 @@ class CacheRequestService extends Component
         $request = Craft::$app->getRequest();
 
         if ($this->getIsStaticInclude()) {
-            // Record the related cache, so we can purge it whenever necessary
-            Blitz::$plugin->generateCache->relatedUri = substr(Craft::$app->getRequest()->url, 1);
+            if (Blitz::$plugin->settings->ssiEnabled) {
+                // Record the related cache, so we can purge it whenever necessary
+                Blitz::$plugin->generateCache->relatedUri = substr(Craft::$app->getRequest()->url, 1);
+            }
 
             return new SiteUriModel([
                 'siteId' => $request->getParam('siteId'),