+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ## 4.3.0 - Unreleased
 ### Added
-- Added a new `ssiEnabled` config setting, that determines whether Blitz will include templates using Server Side Includes (SSI), which must be enabled on the server.
-- Added the `craft.blitz.staticInclude()` and `craft.blitz.dynamicInclude()` template variables, which includes a template using SSI if enabled, otherwise via an AJAX request. The `staticInclude()` method returns a cached result (if one exists), whereas the `dynamicInclude()` method always returns a freshly rendered template.
+- Added a new `ssiEnabled` config setting that enables Blitz to include templates using Server Side Includes (SSI), which must be enabled on the server.
+- Added the `craft.blitz.staticInclude()` template variable, that includes a template using SSI if enabled, otherwise via an AJAX request. The `staticInclude()` method returns a cached result (if one exists).
+- Added the `craft.blitz.dynamicInclude()` template variable, that includes a template via an AJAX request. The `dynamicInclude()` method always returns a freshly rendered template.
 - Added the `blitz/templates/static-include` and `blitz/templates/dynamic-include` controller actions.
 
 ### Deprecated